Fix bug where negative numbers were not being represented correctly
It amazes me this bug has survived this long. Just shows how sparse my testing
really is

+++ src/core/object/gc_ptr.rs
@@ -79,7 +79,7 @@ impl<T> Gc<T> {
     }
 
     fn untag(self) -> (*const u8, Tag) {
-        let ptr = Strict::map_addr(self.ptr, |x| x >> 8);
+        let ptr = Strict::map_addr(self.ptr, |x| ((x as isize) >> 8) as usize);
         let tag = self.tag();
         (ptr, tag)
     }
